Fix scrolling the entire page when comments load
YouTube loads new comments when an over-scroll event is triggered on  #secondary-inner. Unfortunately this over-scroll event requires scrolling the entire page if #secondary manages scrolling.
This fix moves the scrolling behavior to the #secondary container. Over-scrolling on #secondary is also disabled to let the comments load in #secondary-inner without scrolling the entire page.